+++ src/errors.rs
@@ -1,4 +1,4 @@
-use crate::docker::remote;
+use crate::docker;
 use crate::temp;
 
 use std::sync::atomic::{AtomicBool, Ordering};
@@ -102,7 +102,7 @@ unsafe fn termination_handler() {
     // however, we'd need to store the engine path and the argument list as
     // a global CString and `Vec<CString>`, respectively. this atomic guard
     // makes this safe regardless.
-    remote::CONTAINER = None;
+    docker::CONTAINER.terminate();
 
     // EOWNERDEAD, seems to be the same on linux, macos, and bash on windows.
     std::process::exit(130);
